src/backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
-e git+https://github.com/icdev2dev/betaassi@e71013cd9223ba6a7a7e305fa4a64153b60769ae#egg=betaassi
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.35.10
pydantic==2.7.4
pydantic_core==2.18.4
python-engineio==4.9.1
python-socketio==5.11.3
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
